# backend/src/api/routes/sep24.route.ts
import { Router, Request, Response } from 'express';
import { randomUUID } from 'crypto';
import {
  createWithdrawInteractiveUrl,
  createDepositInteractiveUrl,
  isSupportedAsset,
  normalizeAssetCode,
  SUPPORTED_ASSETS,
} from '../../services/kyc.service';
import prisma from '../../lib/prisma';
import { isValidStellarPublicKey } from '../../utils/stellar-address';
import { sep24MetricsMiddleware } from '../middleware/sep24-metrics.middleware';

const router = Router();

router.use(sep24MetricsMiddleware);

interface InteractiveRequest {
  asset_code: string;
  account?: string;
  amount?: string;
  lang?: string;
  quote_id?: string;
  redirect_url?: string;
  on_change_callback?: string;
}

interface InteractiveResponse {
  type: 'interactive_customer_info_needed';
  url: string;
  id: string;
}

const unsupportedAssetResponse = (assetCode: string) => ({
  error: `Asset ${assetCode} is not supported. Supported assets: ${SUPPORTED_ASSETS.join(', ')}`,
});

const invalidAccountResponse = () => ({
  error: 'account must be a valid Stellar public key',
});

const getBaseInteractiveUrl = (): string => process.env.INTERACTIVE_URL || 'http://localhost:3000';

const hasInvalidAccount = (account: unknown): boolean =>
  account !== undefined && !isValidStellarPublicKey(account);

/**
 * @swagger
 * /sep24/transactions/deposit/interactive:
 *   post:
 *     summary: Interactive Deposit
 *     description: SEP-24 Interactive Deposit Endpoint. Returns a URL for the user to complete KYC/Deposit.
 *     tags: [SEP-24]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - asset_code
 *             properties:
 *               asset_code:
 *                 type: string
 *                 description: Asset code to deposit (e.g., USDC, USD, BTC, ETH)
 *                 example: USDC
 *               account:
 *                 type: string
 *                 description: Stellar Ed25519 public key (G...)
 *               amount:
 *                 type: string
 *                 description: Amount to deposit
 *               lang:
 *                 type: string
 *                 description: Language preference for the UI
 *                 default: en
 *     responses:
 *       200:
 *         description: Interactive deposit URL generated
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 type:
 *                   type: string
 *                   example: interactive_customer_info_needed
 *                 url:
 *                   type: string
 *                   description: URL for user to complete deposit
 *                 id:
 *                   type: string
 *                   description: Unique transaction identifier
 *       400:
 *         description: Invalid request parameters
 */
router.post('/transactions/deposit/interactive', async (req: Request, res: Response) => {
  const { asset_code, account, amount, lang = 'en', quote_id, redirect_url, on_change_callback }: InteractiveRequest = req.body;

  const allowedDomains = process.env.SEP24_ALLOWED_CALLBACK_DOMAINS
    ? process.env.SEP24_ALLOWED_CALLBACK_DOMAINS.split(',').filter(Boolean)
    : [];

  if (redirect_url && allowedDomains.length > 0 && !Sep24Service.validateCallbackUrl(redirect_url, allowedDomains)) {
    return res.status(400).json({ error: 'invalid redirect_url domain' });
  }

  if (on_change_callback && allowedDomains.length > 0 && !Sep24Service.validateCallbackUrl(on_change_callback, allowedDomains)) {
    return res.status(400).json({ error: 'invalid on_change_callback domain' });
  }

  if (!asset_code) {
    return res.status(400).json({
      error: 'asset_code is required',
    });
  }

  const normalizedAssetCode = normalizeAssetCode(asset_code);
  if (!isSupportedAsset(normalizedAssetCode)) {
    return res.status(400).json(unsupportedAssetResponse(asset_code));
  }

  if (hasInvalidAccount(account)) {
    return res.status(400).json(invalidAccountResponse());
  }

  if (quote_id) {
    const quote = await prisma.quote.findUnique({ where: { id: quote_id } });
    if (!quote) {
      return res.status(400).json({ error: 'Quote not found' });
    }
    if (quote.expiresAt && new Date() > quote.expiresAt) {
      return res.status(400).json({ error: 'Quote has expired' });
    }
  }

  const transactionId = randomUUID();
  const response: InteractiveResponse = {
    type: 'interactive_customer_info_needed',
    url: createDepositInteractiveUrl({
      baseUrl: getBaseInteractiveUrl(),
      transactionId,
      assetCode: normalizedAssetCode,
      account,
      amount,
      lang,
    }),
    id: transactionId,
  };

  return res.json(response);
});

/**
 * @swagger
 * /sep24/transactions/withdraw/interactive:
 *   post:
 *     summary: Interactive Withdrawal
 *     description: SEP-24 Interactive Withdraw Endpoint. Returns a URL for the user to complete KYC/Withdraw.
 *     tags: [SEP-24]
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - asset_code
 *             properties:
 *               asset_code:
 *                 type: string
 *                 description: Asset code to withdraw (e.g., USDC, USD, BTC, ETH)
 *                 example: USDC
 *               account:
 *                 type: string
 *                 description: Destination Stellar Ed25519 public key (G...)
 *               amount:
 *                 type: string
 *                 description: Amount to withdraw
 *               lang:
 *                 type: string
 *                 description: Language preference for the UI
 *                 default: en
 *     responses:
 *       200:
 *         description: Interactive withdrawal URL generated
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 type:
 *                   type: string
 *                   example: interactive_customer_info_needed
 *                 url:
 *                   type: string
 *                   description: URL for user to complete withdrawal
 *                 id:
 *                   type: string
 *                   description: Unique transaction identifier
 *       400:
 *         description: Invalid request parameters
 */
router.post('/transactions/withdraw/interactive', async (req: Request, res: Response) => {
  const { asset_code, account, amount, lang = 'en', quote_id, redirect_url, on_change_callback }: InteractiveRequest = req.body;

  const allowedDomains = process.env.SEP24_ALLOWED_CALLBACK_DOMAINS
    ? process.env.SEP24_ALLOWED_CALLBACK_DOMAINS.split(',').filter(Boolean)
    : [];

  if (redirect_url && allowedDomains.length > 0 && !Sep24Service.validateCallbackUrl(redirect_url, allowedDomains)) {
    return res.status(400).json({ error: 'invalid redirect_url domain' });
  }

  if (on_change_callback && allowedDomains.length > 0 && !Sep24Service.validateCallbackUrl(on_change_callback, allowedDomains)) {
    return res.status(400).json({ error: 'invalid on_change_callback domain' });
  }

  if (!asset_code) {
    return res.status(400).json({
      error: 'asset_code is required',
    });
  }

  const normalizedAssetCode = normalizeAssetCode(asset_code);
  if (!isSupportedAsset(normalizedAssetCode)) {
    return res.status(400).json(unsupportedAssetResponse(asset_code));
  }

  if (hasInvalidAccount(account)) {
    return res.status(400).json(invalidAccountResponse());
  }

  if (quote_id) {
    const quote = await prisma.quote.findUnique({ where: { id: quote_id } });
    if (!quote) {
      return res.status(400).json({ error: 'Quote not found' });
    }
    if (quote.expiresAt && new Date() > quote.expiresAt) {
      return res.status(400).json({ error: 'Quote has expired' });
    }
  }

  const transactionId = randomUUID();
  const response: InteractiveResponse = {
    type: 'interactive_customer_info_needed',
    url: createWithdrawInteractiveUrl({
      baseUrl: getBaseInteractiveUrl(),
      transactionId,
      assetCode: normalizedAssetCode,
      account,
      amount,
      lang,
    }),
    id: transactionId,
  };

  return res.json(response);
});

export default router;