# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AnchorPoint API",
    "version": "1.0.0",
    "description": "AnchorPoint Backend API for Stellar SEP-24 Anchor Operations",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "AnchorPoint Support",
      "url": "https://github.com/ceejaylaboratory/AnchorPoint"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authorization token obtained from /auth/token endpoint"
      }
    },
...
  },
...
  "paths": {
...
    "/sep24/transactions/deposit/interactive": {
      "post": {
        "summary": "Interactive Deposit",
        "description": "SEP-24 Interactive Deposit Endpoint. Returns a URL for the user to complete KYC/Deposit.",
        "tags": [
          "SEP-24"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "asset_code"
                ],
                "properties": {
                  "asset_code": {
                    "type": "string",
                    "description": "Asset code to deposit (e.g., USDC, USD, BTC, ETH)",
                    "example": "USDC"
                  },
                  "account": {
                    "type": "string",
                    "description": "Stellar Ed25519 public key (G...)"
                  },
                  "amount": {
                    "type": "string",
                    "description": "Amount to deposit"
                  },
                  "lang": {
                    "type": "string",
                    "description": "Language preference for the UI",
                    "default": "en"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interactive deposit URL generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "interactive_customer_info_needed"
                    },
                    "url": {
                      "type": "string",
                      "description": "URL for user to complete deposit"
                    },
                    "id": {
                      "type": "string",
                      "description": "Unique transaction identifier"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          }
        }
      }
    },
    "/sep24/transactions/withdraw/interactive": {
      "post": {
        "summary": "Interactive Withdrawal",
        "description": "SEP-24 Interactive Withdraw Endpoint. Returns a URL for the user to complete KYC/Withdraw.",
        "tags": [
          "SEP-24"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "asset_code"
                ],
                "properties": {
                  "asset_code": {
                    "type": "string",
                    "description": "Asset code to withdraw (e.g., USDC, USD, BTC, ETH)",
                    "example": "USDC"
                  },
                  "account": {
                    "type": "string",
                    "description": "Destination Stellar Ed25519 public key (G...)"
                  },
                  "amount": {
                    "type": "string",
                    "description": "Amount to withdraw"
                  },
                  "lang": {
                    "type": "string",
                    "description": "Language preference for the UI",
                    "default": "en"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interactive withdrawal URL generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "interactive_customer_info_needed"
                    },
                    "url": {
                      "type": "string",
                      "description": "URL for user to complete withdrawal"
                    },
                    "id": {
                      "type": "string",
                      "description": "Unique transaction identifier"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          }
        }
      }
    },
...
  }
}